ormlite

+++ ExemploListView/app/src/main/java/com/example/exemplolistview/control/MainControl.java
@@ -10,65 +10,79 @@ import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.TextView;
 
-import androidx.annotation.NonNull;
-
 import com.example.exemplolistview.R;
+import com.example.exemplolistview.dao.db.EstadoDao;
 import com.example.exemplolistview.model.Estado;
+import com.j256.ormlite.dao.Dao;
 
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+
 
 public class MainControl {
     private Activity activity;
     private EditText editNome;
     private EditText editSigla;
     private TextView tvContagem;
+
     private ListView lvEstado;
     private List<Estado> listEstado;
     private ArrayAdapter<Estado> adapterEstado;
+
     private Estado estado;
+
+    private EstadoDao estadoDao;
 
     public MainControl(Activity activity) {
         this.activity = activity;
+        estadoDao = new EstadoDao(activity);
+
         initComponents();
     }
 
-    private void initComponents() {
-        this.editNome = activity.findViewById(R.id.editNomeEstado);
-        this.editSigla = activity.findViewById(R.id.editSiglaEstado);
+    private void initComponents(){
+        editNome = activity.findViewById(R.id.editNomeEstado);
+        editSigla = activity.findViewById(R.id.editSiglaEstado);
         tvContagem = activity.findViewById(R.id.tvContagem);
         lvEstado = activity.findViewById(R.id.lvEstado);
         configListView();
     }
 
-    private void configListView() {
-        this.listEstado = new ArrayList<>();
-        listEstado.add(new Estado("Acre", "AC"));
-        adapterEstado = new ArrayAdapter<>(activity, android.R.layout.simple_list_item_1);
+    private void configListView(){
+        try {
+            listEstado = estadoDao.getDao().queryForAll();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        adapterEstado = new ArrayAdapter<>(
+                activity,
+                android.R.layout.simple_list_item_1,
+                listEstado
+        );
         lvEstado.setAdapter(adapterEstado);
-        adapterEstado.add(new Estado("Santa Catarina", "SC"));
         cliqueCurto();
         cliqueLongo();
+        atualizarContador();
     }
 
-    private void cliqueCurto() {
+    private void cliqueCurto(){
         lvEstado.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                estado = adapterEstado.getItem(i);
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                estado = adapterEstado.getItem(position);
                 dialogEditarEstado(estado);
             }
         });
     }
-
-    private void cliqueLongo() {
+    private void cliqueLongo(){
         lvEstado.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
             @Override
-            public boolean onItemLongClick(AdapterView<?> adapterView, View view, int i, long l) {
-                estado = adapterEstado.getItem(i);
+            public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
+                estado = adapterEstado.getItem(position);
                 dialogExcluirEstado(estado);
-                //return false; //executa o clique curto junto
-                return true; //executa somente clique longo
+//return false; //Executa o clique curto junto
+                return true; //Executa somente clique longo
             }
         });
     }
@@ -94,7 +108,7 @@ public class MainControl {
         return e;
     }
 
-    private void carregarForm(Estado e) {
+    private void carregarForm(Estado e){
         editNome.setText(e.getNome());
         editSigla.setText(e.getSigla());
     }
@@ -109,15 +123,22 @@ public class MainControl {
         alerta.setMessage(e.toString());
         alerta.setNegativeButton("Fechar", new DialogInterface.OnClickListener() {
             @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
+            public void onClick(DialogInterface dialog, int which) {
                 estado = null;
             }
         });
         alerta.setPositiveButton("Confirmar", new DialogInterface.OnClickListener() {
             @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                excluirEstadoLv(e);
-                atualizarContador();
+            public void onClick(DialogInterface dialog, int which) {
+                try {
+                    if(estadoDao.getDao().delete(estado)>0){
+                        excluirEstadoLv(e);
+                        atualizarContador();
+                    }
+                } catch (SQLException e1) {
+                    e1.printStackTrace();
+                }
+
                 estado = null;
             }
         });
@@ -130,13 +151,13 @@ public class MainControl {
         alerta.setMessage(e.toString());
         alerta.setNegativeButton("Fechar", new DialogInterface.OnClickListener() {
             @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
+            public void onClick(DialogInterface dialog, int which) {
                 estado = null;
             }
         });
         alerta.setPositiveButton("Editar", new DialogInterface.OnClickListener() {
             @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
+            public void onClick(DialogInterface dialog, int which) {
                 carregarForm(e);
             }
         });
@@ -144,18 +165,52 @@ public class MainControl {
     }
 
     public void salvarAction(){
-        if (estado == null){
-            Estado e = getDadosForm();
-            atualizarContador();
-            addEstadoLv(e);
+        if(estado==null){
+            estado = getDadosForm();
         } else {
             Estado e = getDadosForm();
-            atualizarEstadoLv(e);
+            estado.setNome(e.getNome());
+            estado.setSigla(e.getSigla());
         }
+
+        Dao.CreateOrUpdateStatus res;
+        try {
+            res = estadoDao.getDao().createOrUpdate(estado);
+            if(res.isCreated()){
+                addEstadoLv(estado);
+                atualizarContador();
+            } else if(res.isUpdated()){
+                atualizarEstadoLv(estado);
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+
+/*
+if(estado==null){
+Estado e = getDadosForm();
+try {
+if(estadoDao.getDao().create(e)>0){
+addEstadoLv(e);
+atualizarContador();
+}
+} catch (SQLException e1) {
+e1.printStackTrace();
+}
+
+} else {
+Estado e = getDadosForm();
+estado.setNome(e.getNome());
+estado.setSigla(e.getSigla());
+try {
+if(estadoDao.getDao().update(estado)>0){
+atualizarEstadoLv(e);
+}
+} catch (SQLException e1) {
+e1.printStackTrace();
+}
+}*/
         estado = null;
     }
-
-
-
 
 }
